fix eager load
