+++ server/app/lib/analysis_library/base.rb
