+++ server/app/lib/analysis_library/batch_datapoints.rb
