+++ server/app/lib/analysis_library/core.rb
