+++ server/app/lib/analysis_library/diag.rb
