+++ server/app/lib/analysis_library/doe.rb
