+++ server/app/lib/analysis_library/fast99.rb
@@ -33,8 +33,8 @@
 # EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 # *******************************************************************************
 
-#Monte Carlo Estimation of Sobol’ Indices 
-class AnalysisLibrary::Sobol < AnalysisLibrary::Base
+#Fast99 Screening and Sensitivity method
+class AnalysisLibrary::Fast99 < AnalysisLibrary::Base
   include AnalysisLibrary::R::Core
 
   def initialize(analysis_id, analysis_job_id, options = {})
@@ -46,13 +46,8 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
             output_variables: [],
             problem: {
                 algorithm: {
-                    number_of_samples: 30,
-                    random_seed: 1979,
-                    random_seed2: 1973,
-                    order: 1,
-                    nboot: 0,
-                    conf: 0.95,
-                    type: 'sobol',
+                    n: 66,
+                    M: 4,
                     norm_type: 'minkowski',
                     p_power: 2,
                     debug_messages: 0,
@@ -86,12 +81,12 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
     # create an instance for R
     @r = AnalysisLibrary::Core.initialize_rserve(APP_CONFIG['rserve_hostname'],
                                                  APP_CONFIG['rserve_port'])
-    logger.info 'Setting up R for SOBOL Run'
+    logger.info 'Setting up R for Fast99 Run'
     # Initialize some variables that are in the rescue/ensure blocks
     cluster = nil
     begin
       @r.converse("setwd('#{APP_CONFIG['sim_root_path']}')")
-
+      @r.converse(" print(paste('getwd:',getwd()))")
       # make this a core method
       if !@analysis.problem['algorithm']['seed'].nil? && (@analysis.problem['algorithm']['seed'].is_a? Numeric)
         logger.info "Setting R base random seed to #{@analysis.problem['algorithm']['seed']}"
@@ -100,6 +95,10 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
       # R libraries needed for this algorithm
       @r.converse 'library(rjson)'
       @r.converse 'library(sensitivity)'
+      @r.converse 'library(ggplot2)'
+      @r.converse 'library(cowplot)'
+      @r.converse 'library(ggsci)'
+      @r.converse 'library(reshape2)'
 
       # At this point we should really setup the JSON that can be sent to the worker nodes with everything it needs
       # This would allow us to easily replace the queuing system with rabbit or any other json based versions.
@@ -107,70 +106,60 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
       master_ip = 'localhost'
 
       logger.info("Master ip: #{master_ip}")
-      logger.info('Starting GENOUD Run')
+      logger.info('Starting Fast99 Run')
 
       # Quick preflight check that R, MongoDB, and Rails are working as expected. Checks to make sure
       # that the run flag is true.
 
       # TODO: preflight check -- need to catch this in the analysis module
-      if @analysis.problem['algorithm']['order'].nil? || (@analysis.problem['algorithm']['order']).zero?
-        raise 'Value for order was not set or equal to zero (must be 1 or greater)'
+      if @analysis.problem['algorithm']['n'].nil? || (@analysis.problem['algorithm']['n']).zero?
+        raise 'Value for n was not set or equal to zero (must be 1 or greater)'
       end
 
-      if @analysis.problem['algorithm']['conf'].nil? || (@analysis.problem['algorithm']['conf']).zero?
-        raise 'Value for conf was not set or equal to zero (must be 1 or greater)'
-      end
-
-      # TODO: add test for not "minkowski", "maximum", "euclidean", "binary", "manhattan"
-      # if @analysis.problem['algorithm']['norm_type'] != "minkowski", "maximum", "euclidean", "binary", "manhattan"
-      #  raise "P Norm must be non-negative"
-      # end
-
-      if @analysis.problem['algorithm']['p_power'] <= 0
-        raise 'P Norm must be non-negative'
-      end
-
-      if @analysis.problem['algorithm']['number_of_samples'].nil? || (@analysis.problem['algorithm']['number_of_samples']).zero?
-        raise 'Must have number of samples to discretize the parameter space'
+      if (4*@analysis.problem['algorithm']['M']*@analysis.problem['algorithm']['M']+2) <= (@analysis.problem['algorithm']['n'])
+        logger.info 'Value for n was not > 4*M^2+2, will adjust value'
       end
 
       @analysis.problem['algorithm']['objective_functions'] = [] unless @analysis.problem['algorithm']['objective_functions']
       @analysis.save!
+
+      objtrue = @analysis.output_variables.select {|v| v['objective_function'] == true}
+      ug = objtrue.uniq {|v| v['objective_function_group']}
+      logger.info "Number of objective function groups are #{ug.size}"
+      obj_names = []
+      ug.each do |var|
+        obj_names << var['display_name_short']
+      end
+      logger.info "Objective function names #{obj_names}"
 
       pivot_array = Variable.pivot_array(@analysis.id, @r)
       logger.info "pivot_array: #{pivot_array}"
       selected_variables = Variable.variables(@analysis.id)
       logger.info "Found #{selected_variables.count} variables to perturb"
 
+      var_display_names = []
+      selected_variables.each do |var|
+        var_display_names << var.display_name_short
+      end
+      logger.info "Variable display names #{var_display_names}"
+
       # discretize the variables using the LHS sampling method
       @r.converse("print('starting lhs to get min/max')")
       logger.info 'starting lhs to discretize the variables'
 
       lhs = AnalysisLibrary::R::Lhs.new(@r)
-      logger.info "Setting R base random seed to #{@analysis.problem['random_seed']}"
-      @r.converse("set.seed(#{@analysis.problem['algorithm']['random_seed']})")
-      samples, var_types, mins_maxes, var_names = lhs.sample_all_variables(selected_variables, @analysis.problem['algorithm']['number_of_samples'])
-      logger.info "Setting R base random seed to #{@analysis.problem['random_seed2']}"
-      @r.converse("set.seed(#{@analysis.problem['algorithm']['random_seed2']})")
-      samples2, var_types2, mins_maxes2, var_names2 = lhs.sample_all_variables(selected_variables, @analysis.problem['algorithm']['number_of_samples'])
+      samples, var_types, mins_maxes, var_names = lhs.sample_all_variables(selected_variables, 2 * selected_variables.count)
+
+      # Result of the parameter space will be column vectors of each variable
+      logger.info "Samples are #{samples}"
+      logger.info "mins_maxes: #{mins_maxes}"
+      logger.info "var_names: #{var_names}"
+      logger.info("variable types are #{var_types}")
 
       if samples.empty? || samples.size <= 1
         logger.info 'No variables were passed into the options, therefore exit'
         raise "Must have more than one variable to run algorithm.  Found #{samples.size} variables"
       end
-
-      if var_names.empty? || var_names.empty?
-        logger.info 'No variables were passed into the options, therefore exit'
-        raise "Must have at least one variable to run algorithm.  Found #{var_names.size} variables"
-      end
-
-      # Result of the parameter space will be column vectors of each variable
-      logger.info "Samples are #{samples}"
-      logger.info "Samples2 are #{samples2}"
-      logger.info "mins_maxes: #{mins_maxes}"
-      logger.info "var_names: #{var_names}"
-      logger.info "var_names2: #{var_names2}"
-      logger.info("variable types are #{var_types}")
 
       # Start up the cluster and perform the analysis
       cluster = AnalysisLibrary::R::Cluster.new(@r, @analysis.id)
@@ -196,30 +185,26 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
       end
       if cluster.start(worker_ips)
         logger.info "Cluster Started flag is #{cluster.started}"
-        # gen is the number of generations to calculate
-        # varNo is the number of variables (ncol(vars))
-        # popSize is the number of sample points in the variable (nrow(vars))
+
         # convert to float because the value is normally an integer and rserve/rserve-simpler only handles maxint
         @analysis.problem['algorithm']['failed_f_value'] = @analysis.problem['algorithm']['failed_f_value'].to_f
         @r.command(master_ips: master_ip,
                    ips: worker_ips[:worker_ips].uniq,
                    vars: samples.to_dataframe,
-                   vars2: samples2.to_dataframe,
                    vartypes: var_types,
                    varnames: var_names,
                    mins: mins_maxes[:min],
                    maxes: mins_maxes[:max],
-                   order: @analysis.problem['algorithm']['order'],
-                   nboot: @analysis.problem['algorithm']['nboot'],
-                   type: @analysis.problem['algorithm']['type'],
-                   conf: @analysis.problem['algorithm']['conf'],
+                   n: @analysis.problem['algorithm']['n'],
+                   M: @analysis.problem['algorithm']['M'],
                    normtype: @analysis.problem['algorithm']['norm_type'],
                    ppower: @analysis.problem['algorithm']['p_power'],
                    objfun: @analysis.problem['algorithm']['objective_functions'],
                    debug_messages: @analysis.problem['algorithm']['debug_messages'],
                    failed_f: @analysis.problem['algorithm']['failed_f_value'],
-                   mins: mins_maxes[:min],
-                   maxes: mins_maxes[:max]) do
+                   vardisplaynames: var_display_names, 
+                   objnames: obj_names,
+                   uniquegroups: ug.size) do
           %{
             rails_analysis_id = "#{@analysis.id}"
             rails_sim_root_path = "#{APP_CONFIG['sim_root_path']}"
@@ -231,16 +216,18 @@ class AnalysisLibrary::Sobol < AnalysisLibrary::Base
             rails_root_path = "#{Rails.root}"
             rails_host = "#{APP_CONFIG['os_server_host_url']}"
             r_scripts_path = "#{APP_CONFIG['r_scripts_path']}"
+            whoami <- system('whoami', intern = TRUE)
+            print(paste("Fast99.rb whoami:", whoami))
+            hostname <- system('hostname', intern = TRUE)
+            print(paste("Fast99.rb hostname:", hostname))
             rails_exit_guideline_14 = 0
-            source(paste(r_scripts_path,'/sobol.R',sep=''))
-          }
+            source(paste(r_scripts_path,'/fast99.R',sep=''))
+        }
         end
-        logger.info 'Returned from rserve sobol block'
-        # TODO: find any results of the algorithm and save to the analysis
+        logger.info 'Returned from rserve Fast99 block'
       else
         raise 'could not start the cluster (most likely timed out)'
       end
-
     rescue StandardError, ScriptError, NoMemoryError => e
       log_message = "#{__FILE__} failed with #{e.message}, #{e.backtrace.join("\n")}"
       logger.error log_message
