+++ server/app/lib/analysis_library/gaisl.rb
