+++ server/app/lib/analysis_library/optim.rb
