+++ server/app/lib/analysis_library/r/core.rb
