+++ server/app/lib/analysis_library/sequential_search.rb
