+++ server/app/lib/analysis_library/sobol.rb
@@ -33,8 +33,8 @@
 # EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 # *******************************************************************************
 
-#Morris Screening and Sensitivity method
-class AnalysisLibrary::Morris < AnalysisLibrary::Base
+#Monte Carlo Estimation of Sobol’ Indices 
+class AnalysisLibrary::Sobol < AnalysisLibrary::Base
   include AnalysisLibrary::R::Core
 
   def initialize(analysis_id, analysis_job_id, options = {})
@@ -46,11 +46,13 @@ class AnalysisLibrary::Morris < AnalysisLibrary::Base
             output_variables: [],
             problem: {
                 algorithm: {
-                    r: 10,
-                    r2: 20,
-                    levels: 4,
-                    grid_jump: 2,
-                    type: 'oat',
+                    number_of_samples: 30,
+                    random_seed: 1979,
+                    random_seed2: 1973,
+                    order: 1,
+                    nboot: 0,
+                    conf: 0.95,
+                    type: 'sobol',
                     norm_type: 'minkowski',
                     p_power: 2,
                     debug_messages: 0,
@@ -84,12 +86,12 @@ class AnalysisLibrary::Morris < AnalysisLibrary::Base
     # create an instance for R
     @r = AnalysisLibrary::Core.initialize_rserve(APP_CONFIG['rserve_hostname'],
                                                  APP_CONFIG['rserve_port'])
-    logger.info 'Setting up R for Morris Run'
+    logger.info 'Setting up R for SOBOL Run'
     # Initialize some variables that are in the rescue/ensure blocks
     cluster = nil
     begin
       @r.converse("setwd('#{APP_CONFIG['sim_root_path']}')")
-      @r.converse(" print(paste('getwd:',getwd()))")
+
       # make this a core method
       if !@analysis.problem['algorithm']['seed'].nil? && (@analysis.problem['algorithm']['seed'].is_a? Numeric)
         logger.info "Setting R base random seed to #{@analysis.problem['algorithm']['seed']}"
@@ -105,60 +107,70 @@ class AnalysisLibrary::Morris < AnalysisLibrary::Base
       master_ip = 'localhost'
 
       logger.info("Master ip: #{master_ip}")
-      logger.info('Starting Morris Run')
+      logger.info('Starting GENOUD Run')
 
       # Quick preflight check that R, MongoDB, and Rails are working as expected. Checks to make sure
       # that the run flag is true.
 
       # TODO: preflight check -- need to catch this in the analysis module
-      if @analysis.problem['algorithm']['r'].nil? || (@analysis.problem['algorithm']['r']).zero?
-        raise 'Value for r was not set or equal to zero (must be 1 or greater)'
+      if @analysis.problem['algorithm']['order'].nil? || (@analysis.problem['algorithm']['order']).zero?
+        raise 'Value for order was not set or equal to zero (must be 1 or greater)'
       end
 
-      if @analysis.problem['algorithm']['levels'].nil? || (@analysis.problem['algorithm']['levels']).zero?
-        raise 'Value for levels was not set or equal to zero (must be 1 or greater)'
+      if @analysis.problem['algorithm']['conf'].nil? || (@analysis.problem['algorithm']['conf']).zero?
+        raise 'Value for conf was not set or equal to zero (must be 1 or greater)'
+      end
+
+      # TODO: add test for not "minkowski", "maximum", "euclidean", "binary", "manhattan"
+      # if @analysis.problem['algorithm']['norm_type'] != "minkowski", "maximum", "euclidean", "binary", "manhattan"
+      #  raise "P Norm must be non-negative"
+      # end
+
+      if @analysis.problem['algorithm']['p_power'] <= 0
+        raise 'P Norm must be non-negative'
+      end
+
+      if @analysis.problem['algorithm']['number_of_samples'].nil? || (@analysis.problem['algorithm']['number_of_samples']).zero?
+        raise 'Must have number of samples to discretize the parameter space'
       end
 
       @analysis.problem['algorithm']['objective_functions'] = [] unless @analysis.problem['algorithm']['objective_functions']
       @analysis.save!
-
-      objtrue = @analysis.output_variables.select {|v| v['objective_function'] == true}
-      ug = objtrue.uniq {|v| v['objective_function_group']}
-      logger.info "Number of objective function groups are #{ug.size}"
-      obj_names = []
-      ug.each do |var|
-        obj_names << var['display_name_short']
-      end
-      logger.info "Objective function names #{obj_names}"
 
       pivot_array = Variable.pivot_array(@analysis.id, @r)
       logger.info "pivot_array: #{pivot_array}"
       selected_variables = Variable.variables(@analysis.id)
       logger.info "Found #{selected_variables.count} variables to perturb"
 
-      var_display_names = []
-      selected_variables.each do |var|
-        var_display_names << var.display_name_short
-      end
-      logger.info "Variable display names #{var_display_names}"
-
       # discretize the variables using the LHS sampling method
       @r.converse("print('starting lhs to get min/max')")
       logger.info 'starting lhs to discretize the variables'
 
       lhs = AnalysisLibrary::R::Lhs.new(@r)
-      samples, var_types, mins_maxes, var_names = lhs.sample_all_variables(selected_variables, 2 * selected_variables.count)
-
-      # Result of the parameter space will be column vectors of each variable
-      logger.info "Samples are #{samples}"
-      logger.info "mins_maxes: #{mins_maxes}"
-      logger.info "var_names: #{var_names}"
-      logger.info("variable types are #{var_types}")
+      logger.info "Setting R base random seed to #{@analysis.problem['random_seed']}"
+      @r.converse("set.seed(#{@analysis.problem['algorithm']['random_seed']})")
+      samples, var_types, mins_maxes, var_names = lhs.sample_all_variables(selected_variables, @analysis.problem['algorithm']['number_of_samples'])
+      logger.info "Setting R base random seed to #{@analysis.problem['random_seed2']}"
+      @r.converse("set.seed(#{@analysis.problem['algorithm']['random_seed2']})")
+      samples2, var_types2, mins_maxes2, var_names2 = lhs.sample_all_variables(selected_variables, @analysis.problem['algorithm']['number_of_samples'])
 
       if samples.empty? || samples.size <= 1
         logger.info 'No variables were passed into the options, therefore exit'
         raise "Must have more than one variable to run algorithm.  Found #{samples.size} variables"
       end
+
+      if var_names.empty? || var_names.empty?
+        logger.info 'No variables were passed into the options, therefore exit'
+        raise "Must have at least one variable to run algorithm.  Found #{var_names.size} variables"
+      end
+
+      # Result of the parameter space will be column vectors of each variable
+      logger.info "Samples are #{samples}"
+      logger.info "Samples2 are #{samples2}"
+      logger.info "mins_maxes: #{mins_maxes}"
+      logger.info "var_names: #{var_names}"
+      logger.info "var_names2: #{var_names2}"
+      logger.info("variable types are #{var_types}")
 
       # Start up the cluster and perform the analysis
       cluster = AnalysisLibrary::R::Cluster.new(@r, @analysis.id)
@@ -184,28 +196,28 @@ class AnalysisLibrary::Morris < AnalysisLibrary::Base
       end
       if cluster.start(worker_ips)
         logger.info "Cluster Started flag is #{cluster.started}"
-
+        # gen is the number of generations to calculate
+        # varNo is the number of variables (ncol(vars))
+        # popSize is the number of sample points in the variable (nrow(vars))
         # convert to float because the value is normally an integer and rserve/rserve-simpler only handles maxint
         @analysis.problem['algorithm']['failed_f_value'] = @analysis.problem['algorithm']['failed_f_value'].to_f
         @r.command(master_ips: master_ip,
                    ips: worker_ips[:worker_ips].uniq,
                    vars: samples.to_dataframe,
+                   vars2: samples2.to_dataframe,
                    vartypes: var_types,
                    varnames: var_names,
                    mins: mins_maxes[:min],
                    maxes: mins_maxes[:max],
-                   levels: @analysis.problem['algorithm']['levels'],
-                   r: @analysis.problem['algorithm']['r'],
-                   r2: @analysis.problem['algorithm']['r2'],
+                   order: @analysis.problem['algorithm']['order'],
+                   nboot: @analysis.problem['algorithm']['nboot'],
                    type: @analysis.problem['algorithm']['type'],
-                   grid_jump: @analysis.problem['algorithm']['grid_jump'],
+                   conf: @analysis.problem['algorithm']['conf'],
                    normtype: @analysis.problem['algorithm']['norm_type'],
                    ppower: @analysis.problem['algorithm']['p_power'],
                    objfun: @analysis.problem['algorithm']['objective_functions'],
                    debug_messages: @analysis.problem['algorithm']['debug_messages'],
-                   failed_f: @analysis.problem['algorithm']['failed_f_value'],
-                   vardisplaynames: var_display_names, objnames: obj_names,
-                   mins: mins_maxes[:min], maxes: mins_maxes[:max], uniquegroups: ug.size) do
+                   failed_f: @analysis.problem['algorithm']['failed_f_value']) do
           %{
             rails_analysis_id = "#{@analysis.id}"
             rails_sim_root_path = "#{APP_CONFIG['sim_root_path']}"
@@ -217,18 +229,16 @@ class AnalysisLibrary::Morris < AnalysisLibrary::Base
             rails_root_path = "#{Rails.root}"
             rails_host = "#{APP_CONFIG['os_server_host_url']}"
             r_scripts_path = "#{APP_CONFIG['r_scripts_path']}"
-            whoami <- system('whoami', intern = TRUE)
-            print(paste("Morris.rb whoami:", whoami))
-            hostname <- system('hostname', intern = TRUE)
-            print(paste("Morris.rb hostname:", hostname))
             rails_exit_guideline_14 = 0
-            source(paste(r_scripts_path,'/morris.R',sep=''))
-        }
+            source(paste(r_scripts_path,'/sobol.R',sep=''))
+          }
         end
-        logger.info 'Returned from rserve morris block'
+        logger.info 'Returned from rserve sobol block'
+        # TODO: find any results of the algorithm and save to the analysis
       else
         raise 'could not start the cluster (most likely timed out)'
       end
+
     rescue StandardError, ScriptError, NoMemoryError => e
       log_message = "#{__FILE__} failed with #{e.message}, #{e.backtrace.join("\n")}"
       logger.error log_message
