+++ server/app/lib/analysis_library/spea_nrel.rb
